backend/services/integrated_investment_service/requirements.txt
yfinance>=0.2.36
google-genai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
